modify: module layout
