+++ crates/voicevox_core/src/engine/mod.rs
@@ -1,1 +1,3 @@
 mod kana_parser;
+mod model;
+mod mora_list;
